backend_django/pyproject.toml
[tool.poetry]
name = "firebasevezba"
version = "0.1.0"
description = ""
authors = ["Ivan"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
firebase-admin = "^6.0.1"
django = "^4.1.5"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
django-environ = "^0.10.0"
celery = "^5.2.7"
redis = "^4.5.4"
aiohttp = "^3.8.4"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
